moved claude plugin script tests into a symmetric directory under a new top-level , that way test files don't ship to our users
