+++ tests/providers/claude/plugin/scripts/cli.test.mjs
