+++ tests/providers/claude/plugin/scripts/hookHelpers.test.mjs
@@ -1,6 +1,6 @@
 import assert from 'node:assert/strict';
 import test from 'node:test';
-import { runHook } from './hookHelpers.mjs';
+import { runHook } from '../../../../../providers/claude/plugin/scripts/hookHelpers.mjs';
 
 test('runHook swallows callback failures without rejecting', async () => {
   await assert.doesNotReject(() =>
