+++ tests/providers/claude/plugin/scripts/transcriptHelpers.test.mjs
